+++ Civi/DataProcessor/DataSpecification/CustomFieldSpecification.php
@@ -35,7 +35,7 @@ class CustomFieldSpecification extends FieldSpecification {
     $this->customFieldTitle = $field['label'];
 
     $options = \CRM_Core_PseudoConstant::get('CRM_Core_BAO_CustomField', 'custom_' . $this->customFieldId, array(), 'search');
-    parent::__construct($field['column_name'], $field['data_type'], $custom_group_title. ': '.$this->customFieldTitle, $options, $alias);
+    parent::__construct($alias, $field['data_type'], $custom_group_title. ': '.$this->customFieldTitle, $options, $alias);
 
   }
 
@@ -58,6 +58,13 @@ class CustomFieldSpecification extends FieldSpecification {
       return false;
     }
     return true;
+  }
+
+  /**
+   * @return String
+   */
+  public function getName() {
+    return $this->customFieldColumnName;
   }
 
 }